+++ MangoAPI.BusinessLogic/ApiCommands/Users/UpdateUserAccountInfoCommandHandler.cs
@@ -36,7 +36,8 @@ namespace MangoAPI.BusinessLogic.ApiCommands.Users
             {
                 var userChats = await _postgresDbContext.UserChats
                     .Include(x => x.Chat)
-                    .Where(x => x.UserId == user.Id && x.Chat.CommunityType == CommunityType.DirectChat)
+                    .Where(x => x.UserId == user.Id && 
+                                x.Chat.CommunityType == (int) CommunityType.DirectChat)
                     .Select(x => x.Chat)
                     .ToListAsync(cancellationToken);
